Fix a bug where 5-card straights error out
The method that checks if five cards are in order (a straight) assumed
that it would be iterating through 7 cards.  In checking for a royal
flush, only 5 cards are checked for a straight so the set of cards was
smaller than the indexes we were using to check, and returned nil, which
resulted in a NoMethodError when getting the card's rank.  This changes
makes it so we check the size of the cards before checking if there's a
straight.